Support adding metadata for trusting executables.
Summary: tl;dr this fixes the need to have to select open or execute when clicking on a .desktop file.

Test Plan: To be fair I'm not sure of the mechanism but somehow we should be able to set some property of our install.desktop so that when you double click on it, it executes rather than prompting.

Reviewers: tsimonq2

Reviewed By: tsimonq2

Differential Revision: https://phab.lubuntu.me/D29